requirements.txt
python-telegram-bot==21.6
tweepy==4.14.0
httpx==0.27.0
APScheduler==3.10.4
Pillow==10.4.0
python-dotenv==1.0.1
